operation-and-maintenance/maintain-incidementmanagement-slack-pagerduty/mcp/incident_management/infrastructure/requirements.txt
# AWS CDK Core
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# CDK Specific Libraries
aws-cdk.aws-lambda-python-alpha>=2.100.0a0

# CDK Nag for security and compliance checks
cdk-nag>=2.27.0

# Python Dependencies
boto3>=1.28.0
botocore>=1.31.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
jinja2>=3.1.2